Adjusted package names and imports for Eclipse usage, moved AlarmService into package sub dir added @SuppressLint("NewApi") to suppress min api version warnings in Eclipse
