+++ src/com/firebirdberlin/nightdream/AlarmClock.java
@@ -30,10 +30,15 @@ import java.text.SimpleDateFormat;
 import java.util.Calendar;
 import java.util.Locale;
 
+import com.firebirdberlin.nightdream.models.SimpleTime;
+
 import static android.text.format.DateFormat.getBestDateTimePattern;
 import static android.text.format.DateFormat.is24HourFormat;
 
-class AlarmClock extends View {
+import android.annotation.SuppressLint;
+
+@SuppressLint("NewApi")
+public class AlarmClock extends View {
     private static String TAG ="NightDream.AlarmClock";
     final private Handler handler = new Handler();
     public boolean isVisible = false;
