+++ src/com/firebirdberlin/nightdream/NightDreamService.java
@@ -1,5 +1,6 @@
 package com.firebirdberlin.nightdream;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
@@ -18,7 +19,14 @@ import android.widget.Toast;
 import de.greenrobot.event.EventBus;
 import java.util.Calendar;
 
+import com.firebirdberlin.nightdream.events.OnClockClicked;
+import com.firebirdberlin.nightdream.events.OnLightSensorValueTimeout;
+import com.firebirdberlin.nightdream.events.OnNewAmbientNoiseValue;
+import com.firebirdberlin.nightdream.events.OnNewLightSensorValue;
+import com.firebirdberlin.nightdream.ui.NightDreamUI;
 
+
+@SuppressLint("NewApi")
 public class NightDreamService extends DreamService implements View.OnTouchListener {
 
     TextView current;
