+++ src/com/firebirdberlin/nightdream/PowerConnectionReceiver.java
@@ -1,5 +1,6 @@
 package com.firebirdberlin.nightdream;
 
+import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.app.AlarmManager;
 import android.app.PendingIntent;
@@ -12,6 +13,8 @@ import android.os.Build;
 import android.os.Bundle;
 import java.util.Calendar;
 import java.util.GregorianCalendar;
+
+import com.firebirdberlin.nightdream.models.SimpleTime;
 
 public class PowerConnectionReceiver extends BroadcastReceiver {
     private static int PENDING_INTENT_START_APP = 0;
@@ -67,7 +70,8 @@ public class PowerConnectionReceiver extends BroadcastReceiver {
         return false;
     }
 
-    static public void schedule(Context context) {
+    @SuppressLint("NewApi")
+	static public void schedule(Context context) {
         Intent alarmIntent = new Intent(context, PowerConnectionReceiver.class);
         PendingIntent pendingIntent = PendingIntent.getBroadcast(context, PENDING_INTENT_START_APP,
                                                                  alarmIntent,
